Calculate amount and return txAuthor
